+++ src/main/java/org/spongepowered/api/event/ExperienceEvent.java
@@ -31,7 +31,7 @@ public interface ExperienceEvent {
      *
      * @return The experience to receive
      */
-    int getExp();
+    double getExp();
 
     /**
      * Set the amount of experience after an event has been processed.
@@ -39,6 +39,6 @@ public interface ExperienceEvent {
      *
      * @param exp The experience to give or take
      */
-    void setExp(int exp);
+    void setExp(double exp);
 
 }
